+++ CHANGELOG.markdown
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 
 ### Fixed
 
+* Fixed constants being always created in the global scope ([#352](https://github.com/cardillan/mindcode/issues/352)).
 * Fixed and reactivated the `op floor` removal optimization after it has been deactivated in 3.17.4.
 * Fixed the `error()` function being compiled even when the `error-function` option was set to `false`.
 * Fixed the `assertbounds` instruction not being properly handled when the [implicit integer conversions optimization](/doc/syntax/optimizations/EXPRESSION-OPTIMIZATION.markdown#handling-non-negative-implicit-integer-conversions) was applied, potentially leading to compiler errors or mistakenly reported runtime errors.